// frontend/frontend/app/src/app/model/userModel.ts
import { Type } from "@angular/core";

export class User{
    username: string;
    password: string;
    index: string;
    studiesType:string;
    name: string;
    surname: string;
    status: string;
    numOfLogin: number;
    employeeFlag: number;
    adress:string;
    typeOf: string;
    roomNumber: string;
    personalData: string;
    mobilePhone: string;
    subjects: Array<Object>;
    mail: string;
    profession:string;
    website:string;
    picture:string;

}